+++ algorithms/flybyPoint/flybyPointAlgorithm.h
@@ -10,9 +10,9 @@
 
 /*! structure containing the attitude guidance outputs of the algorithm */
 struct FlybyPointOutput {
-    Eigen::Vector3d sigma_RN = Eigen::Vector3d::Zero();
-    Eigen::Vector3d omega_RN_N = Eigen::Vector3d::Zero();
-    Eigen::Vector3d domega_RN_N = Eigen::Vector3d::Zero();
+    Eigen::Vector3f sigma_RN = Eigen::Vector3f::Zero();
+    Eigen::Vector3f omega_RN_N = Eigen::Vector3f::Zero();
+    Eigen::Vector3f domega_RN_N = Eigen::Vector3f::Zero();
     bool collinearityTrigger = false;
     bool maxRateTrigger = false;
     bool maxAccelerationTrigger = false;
